% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Tidy NOAA}
\usage{
eq_clean_data(rawNOAA)
}
\arguments{
\item{rawNOAA}{}
}
\value{
a data.frame with LNG/LAT in double, date column as Date object
}
\description{
this function is the utility responsabile for clean up the raw data file into a specifically designed data.frame
for fourther
}
\examples{
eq_clean_data(readRawNOAA(filetabdelimited))
}
