% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(rawNOAA)
}
\arguments{
\item{rawNOAA}{}
}
\value{
the original data.frame with LOCATION_NAME mutated
}
\description{
this function take in the NOAA data.frame and process the column LOCATION_NAME.
The column will get the first country removed and convert the rest to mixed case. IE, the value "ITALY: SICILY" will
be converted in "Sicily".
}
\examples{
eq_location_clean(rawNOAA)

}
