% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{date

color and size, as inerited by Point}

\item{data}{a data.frame as returned by storm_observation}

\item{...}{}
}
\value{
a plot
}
\description{
this function create the geom_timeline for a given NOAA cleaned data.frame
}
\examples{

}
