% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomTimeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeline}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 6.}
\usage{
GeomTimeline
}
\description{
this geom is used by the function geom_Timeline and shuld not be used directly
}
\keyword{datasets}
