% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{readRawNOAA}
\alias{readRawNOAA}
\title{Title}
\usage{
readRawNOAA(filename)
}
\arguments{
\item{filename}{}
}
\value{
a data.frame with raw NOAA data
}
\description{
Title
}
\examples{
readRawNOAA(file.path('~',"signif.txt"))
}
