% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{titolize}
\alias{titolize}
\title{titolize}
\usage{
titolize(x)
}
\arguments{
\item{x}{characters, or list of characters.}
}
\value{
a mixed case string, or a list of mixed case sting if x was a list
}
\description{
This function make every first letter in uppercase of a string, function is vectorized.
}
\examples{
titolize("my string is lowercase")
}
